Fixed fadeDist bugging out on low values
Setting fadeDist to 1 in the UI can effectively brick the UI as the player position refers to the avatar root, not the head. This can make it impossible to effectively reduce the distance to less than 1 and may require rejoining the world to fix.

This change compares against the actual head position instead so that the user can intuitively move closer to the UI and make it functional again.